chore: refactor into small components
