+++ src/components/App/App.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import TopBar from '../TopBar/TopBar';
+import HomePage from '../HomePage/HomePage';
+
+export default function App() {
+  return (
+    <>
+      <TopBar />
+      <HomePage />
+    </>
+  );
+}
