+++ src/components/HomePage/HomePage.css
@@ -3,22 +3,10 @@
   padding: 10px;
 }
 
-.title {
-  font-family: "Public Sans", sans-serif;
-  font-weight: 800;
-}
-
 .links {
   font-size: 70%;
   font-family: "Public Sans", sans-serif;
   font-weight: 600;
-}
-
-.App-header {
-  background-color: #fcf4e8;
-  border-bottom: solid 2px #000000;
-  font-weight: 500;
-  color: #c97e2e !important;
 }
 
 .full-width-image {
@@ -65,12 +53,14 @@
   letter-spacing: 1px;
   word-spacing: 1.25%;
 }
+
 .p2 {
   font-size: 1.5vw;
   text-align: center;
   font-weight: 500;
   color: #3f342c;
 }
+
 .p3 {
   font-size: 2vw;
   font-family: "Public Sans", sans-serif;
@@ -79,6 +69,7 @@
   color: #e68a00;
   margin: 4%;
 }
+
 .p4 {
   font-size: 45px;
 
@@ -86,6 +77,7 @@
   text-align: center;
   margin: 0%;
 }
+
 .p5 {
   font-size: 45px;
   text-align: center;
@@ -136,7 +128,7 @@ p {
 .footer {
   padding-top: 3% !important;
   padding-left: 5% !important;
-  color: #ffffff;
+  color: #fff;
   background-color: #3c2f29;
   min-height: 20vh;
   font-size: 2.1vw;
@@ -145,11 +137,6 @@ p {
 
 .h-text {
   transform: rotate(-180deg);
-}
-
-p {
-  word-spacing: 1%;
-  line-height: 1.25;
 }
 
 .nopadding-left {
@@ -171,7 +158,8 @@ p {
   margin-left: 4rem;
   margin-top: 19rem;
   font-weight: 500;
-  /* display-direction: column;*/
+
+  /* display-direction: column; */
 }
 
 #kevinquote {
@@ -182,11 +170,13 @@ p {
   position: absolute;
   color: #000;
   font-size: 1rem;
+
   /* margin-left: 2rem; */
   margin-left: 16rem;
   margin-top: 18rem;
+
   /* margin: 16rem;
-  margin-left: 13rem; */
+    margin-left: 13rem; */
   font-weight: 600;
 }
 
@@ -203,16 +193,15 @@ p {
 #speechBubble {
   margin-bottom: 2px;
 }
+
 p.bubble {
   position: absolute;
   width: 18rem;
   text-align: center;
   line-height: 1.4em;
   margin: 0 0.1rem;
-
   border: 5px solid #3a2419;
   border-radius: 0.4rem;
-
   padding: 1px;
   font-size: 1.7rem;
   font-weight: 600;
@@ -220,25 +209,28 @@ p.bubble {
   color: #fff;
 }
 
-p.bubble:before,
-p.bubble:after {
+p.bubble::before,
+p.bubble::after {
   content: " ";
   position: absolute;
   width: 0;
   height: 0;
 }
-p.speech:before {
+
+p.speech::before {
   left: 4px;
   bottom: -47px;
   border: 24px solid;
   border-color: #3a2419 transparent transparent #3a2419;
 }
-p.speech:after {
+
+p.speech::after {
   left: 24px;
   bottom: -27px;
   border: 5px solid;
   border-color: #3a2419 transparent transparent #3a2419;
 }
+
 .kevinquote1,
 .kevinquote2 {
   position: absolute;
@@ -246,10 +238,12 @@ p.speech:after {
   font-weight: 700;
   font-size: 1.3rem;
 }
+
 .kevinquote1 {
   margin-left: 5rem;
   margin-top: 8rem;
 }
+
 .kevinquote2 {
   margin-left: 3.5rem;
   margin-top: 10rem;
@@ -259,6 +253,7 @@ p.speech:after {
   font-weight: 600;
   margin-bottom: 1px;
 }
+
 #quoteunderline {
   margin-top: 1rem;
   text-decoration: underline solid rgb(148, 110, 15) 0.53rem;
@@ -267,19 +262,21 @@ p.speech:after {
 /* media  */
 
 /* iphone 7 plus  */
-@media only screen and (min-width: 0px) and (max-width: 480px) {
+@media only screen and (min-width: 0) and (max-width: 480px) {
   p.bubble {
     width: 6rem;
     padding: 1px;
     font-size: 0.6rem;
   }
-  p.speech:before {
+
+  p.speech::before {
     left: 1px;
     bottom: -15px;
     border: 8px solid;
     border-color: #3a2419 transparent transparent #3a2419;
   }
-  p.speech:after {
+
+  p.speech::after {
     left: 1px;
     bottom: -3px;
     border: 4px solid;
@@ -301,6 +298,7 @@ p.speech:after {
     margin-top: 4rem;
     margin-left: 0.5rem;
   }
+
   .kevinquote2 {
     margin-top: 4.7rem;
     margin-left: 0.05rem;
@@ -327,18 +325,20 @@ p.speech:after {
     padding-right: 1px;
   }
 
-  p.speech:before {
+  p.speech::before {
     left: 4px;
     bottom: -25px;
     border: 10px solid;
     border-color: #3a2419 transparent transparent #3a2419;
   }
-  p.speech:after {
+
+  p.speech::after {
     left: 9px;
     bottom: -8px;
     border: 3px solid;
     border-color: #3a2419 transparent transparent #3a2419;
   }
+
   .kevinquote1,
   .kevinquote2 {
     font-size: 0.8rem;
@@ -348,21 +348,25 @@ p.speech:after {
     margin-top: 4rem;
     margin-left: 1rem;
   }
+
   .kevinquote2 {
     margin-top: 4.8rem;
     margin-left: 0.1rem;
   }
+
   .kevimage img {
     height: auto;
     width: 15rem;
     margin-left: 4rem;
     z-index: -1;
   }
+
   .yearquote p {
     font-size: 1.5em;
     margin-left: 0.1rem;
     margin-top: 13rem;
   }
+
   #kevruddName {
     margin-top: 13rem;
     margin-left: 7rem;
@@ -375,13 +379,15 @@ p.speech:after {
     font-size: 1rem;
     padding: 5px;
   }
-  p.speech:before {
+
+  p.speech::before {
     left: 4px;
     bottom: -40px;
     border: 20px solid;
     border-color: #3a2419 transparent transparent #3a2419;
   }
-  p.speech:after {
+
+  p.speech::after {
     left: 24px;
     bottom: -20px;
     border: 3px solid;
@@ -409,16 +415,12 @@ p.speech:after {
     margin-top: 4rem;
     margin-left: 3rem;
   }
+
   .kevinquote2 {
     margin-top: 5rem;
     margin-left: 2rem;
   }
 
-  .yearquote p {
-    font-size: 1.7em;
-    margin-left: 3rem;
-    margin-top: 12rem;
-  }
   #kevruddName {
     margin-top: 13rem;
   }
