+++ src/components/TopBar/TopBar.css
@@ -1,0 +1,11 @@
+.App-header {
+  background-color: #fcf4e8;
+  border-bottom: solid 2px #000;
+  font-weight: 500;
+  color: #c97e2e !important;
+}
+
+.title {
+  font-family: "Public Sans", sans-serif;
+  font-weight: 800;
+}
