+++ src/components/TopBar/TopBar.test.jsx
@@ -1,0 +1,7 @@
+import { render } from '@testing-library/react';
+import TopBar from './TopBar';
+
+test('renders TopBar', () => {
+  const { getByText } = render(<TopBar />);
+  getByText('SeeChange.');
+});
